+++ src/fhir/mod.rs
@@ -1,0 +1,7 @@
+pub mod code;
+pub mod condition;
+pub mod encounter;
+pub mod observation;
+pub mod patient;
+pub mod procedure;
+pub mod resources;
